Split calling backend into protocol and separate implementations
