+++ src/yoyopod/communication/calling/voice_notes.py
@@ -11,7 +11,7 @@ from typing import Callable
 
 from loguru import logger
 
-from yoyopod.communication.calling.backend import VoIPBackend
+from yoyopod.communication.calling.backend_protocol import VoIPBackend
 from yoyopod.communication.messaging import VoIPMessageStore
 from yoyopod.communication.models import (
     MessageDeliveryChanged,
